+++ src/python/impactx/dashboard/Analyze/plot_PhaseSpaceProjections/phaseSpace.py
@@ -15,11 +15,11 @@ import importlib
 
 from impactx import ImpactX
 
-from ...Input.distributionParametersCard.distributionMain import (
+from ...Input.distributionParameters.distributionMain import (
     save_distribution_parameters,
 )
-from ...Input.latticeConfigurationCard.latticeMain import save_lattice_elements
-from ..plot_phase_space.phaseSpaceSettings import adjusted_settings_plot
+from ...Input.latticeConfiguration.latticeMain import save_lattice_elements
+from ..plot_PhaseSpaceProjections.phaseSpaceSettings import adjusted_settings_plot
 
 def run_simulation(save_png=True):
     """
